pyproject.toml
[tool.poetry]
name = "infomfss"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
generics = "^7.0.0"
coverage = "^7.6.4"
coreapi = "^2.3.3"
requests = "^2.32.3"
pyopenssl = "^24.2.1"
eventlet = "^0.37.0"
django-environ = "^0.11.2"
libsass = "^0.23.0"
django-compressor = "^4.5.1"
django-sass-processor = "^1.4.1"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-smart-selects = "^1.6.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
